+++ .env.example
@@ -1,0 +1,9 @@
+# common variables
+CHAIN_ID="" # mainnet: 1, goerli: 5
+FRONTEND_URL=""
+OWNER_PRIVATE_KEY=""
+RPC_PROVIDER_URL=""
+TOKEN_ADDRESS="" # // DAI address, mainnet: 0x6b175474e89094c44da98b954eedeac495271d0f, goerli: 0x11fE4B6AE13d2a6055C8D9cF65c55bac32B5d844
+# variables depending on spender (bounty hunter)
+AMOUNT_IN_ETH="1" # amount in ether, 1 AMOUNT_IN_ETH = 1000000000000000000 WEI
+SPENDER_ADDRESS=""
