+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@types/node": "^18.13.0",
     "@uniswap/permit2-sdk": "^1.2.0",
+    "dotenv": "^16.0.3",
     "ethers": "^5.7.2",
     "prettier": "^2.8.4",
     "tsx": "^3.12.3"
